Refactored out mysql for sqlite
