+++ config.json
@@ -1,9 +1,6 @@
 {
   "PORT": 8000,
-  
-  "SESSION_SECRET": "random-data-to-seed-session-data",
-  "SQL_HOST": "db",
-  "SQL_DATABASE": "jrtechs_blog",
-  "SQL_USER": "root",
-  "SQL_PASSWORD": "password"
+  "CACHE": false,
+  "ADMIN_CHECK": true,
+  "SESSION_SECRET": "random-data-to-seed-session-data"
 }
