+++ docker-compose.yml
@@ -4,14 +4,14 @@ version: '3'
 
 # mysql --port=3306 --host=127.0.0.1 -u root --password=password
 services:
-    db:
-        image: mysql
-        command: --default-authentication-plugin=mysql_native_password
-        restart: always
-        environment:
-            MYSQL_ROOT_PASSWORD: password
-        volumes:
-            - "./db:/var/lib/mysql"
+    # db:
+    #     image: mysql
+    #     command: --default-authentication-plugin=mysql_native_password
+    #     restart: always
+    #     environment:
+    #         MYSQL_ROOT_PASSWORD: password
+    #     volumes:
+    #         - "./db:/var/lib/mysql"
 
     blog:
         build: .
@@ -20,5 +20,5 @@ services:
         ports:
             - 8000:8000
         restart: always
-        links:
-            - db:database
+        # links:
+        #     - db:database
