+++ package.json
@@ -19,6 +19,7 @@
     "rss": "^1.2.2",
     "sanitizer": "^0.1.3",
     "sendmail": "^1.4.1",
+    "sqlite3": "^5.0.2",
     "whiskers": "^0.4.0"
   },
   "scripts": {
