+++ routes/index.js
@@ -60,6 +60,7 @@ routes.use('/rss', feed);
 //blog home page
 routes.get('/', (request, result) =>
 {
+    console.log("ummmm ping?");
     pageBuilder.buildBlogPage(request, result, require("../blog/homePage").main)
 });
