+++ sitemap.txt
@@ -1,24 +1,46 @@
 http://jrtechs.net/
+http://jrtechs.net/category/data-science
 http://jrtechs.net/category/hardware
 http://jrtechs.net/category/java
+http://jrtechs.net/category/open-source
 http://jrtechs.net/category/other
+http://jrtechs.net/category/photography
 http://jrtechs.net/category/programming
 http://jrtechs.net/category/projects
 http://jrtechs.net/category/web-development
-http://jrtechs.net/hardware/2018-rochester-maker-faire
-http://jrtechs.net/hardware/ddr4-ram-introduction
-http://jrtechs.net/hardware/hard-drive-speeds
+http://jrtechs.net/data-science/visualizing-fitbit-gps-data
+http://jrtechs.net/data-science/quadtree-animations-with-matplotlib
+http://jrtechs.net/data-science/implementing-a-quadtree-in-python
+http://jrtechs.net/data-science/node2vec-with-steam-data
+http://jrtechs.net/data-science/time-spent-in-steam-games
+http://jrtechs.net/data-science/cuda-vs-cpu-performance
+http://jrtechs.net/data-science/creating-pixel-art-with-open-cv
+http://jrtechs.net/data-science/gans-in-pytorch
+http://jrtechs.net/data-science/pytorch-headfirst
+http://jrtechs.net/data-science/word-embeddings-part-2
+http://jrtechs.net/data-science/word-embeddings
+http://jrtechs.net/data-science/image-clustering-with-k-means
+http://jrtechs.net/data-science/graphing-my-life-with-matplotlib
+http://jrtechs.net/data-science/developing-an-ai-to-play-asteroids-part-1
+http://jrtechs.net/data-science/csci-331-final-review
+http://jrtechs.net/data-science/csci-331-review-2
+http://jrtechs.net/data-science/csci-331-review-1
+http://jrtechs.net/data-science/a-closer-look-at-fitbit-data
+http://jrtechs.net/data-science/r-programming-language
+http://jrtechs.net/data-science/is-using-ml-for-antivirus-safe
+http://jrtechs.net/data-science/lets-build-a-genetic-algorithm
+http://jrtechs.net/java/running-a-minecraft-server-with-docker
+http://jrtechs.net/java/running-scala-code-in-docker
+http://jrtechs.net/java/parallel-java-performance-overview
+http://jrtechs.net/java/fun-with-functional-java
 http://jrtechs.net/java/gremlin-in-10-minutes
 http://jrtechs.net/java/top-three-recommended-java-ides
 http://jrtechs.net/java/bash-usr-bin-java-cannot-execute-binary-file
-http://jrtechs.net/other/college-cookbook
-http://jrtechs.net/other/2018-in-review
-http://jrtechs.net/other/morality-of-self-driving-cars
-http://jrtechs.net/other/my-college-essay
-http://jrtechs.net/other/deepfakes
-http://jrtechs.net/other/why-do-i-blog
-http://jrtechs.net/other/ways-to-avoid-getting-hacked
-http://jrtechs.net/other/should-you-run-a-server-on-desktop-hardware
+http://jrtechs.net/photography/segmenting-images-with-quadtrees
+http://jrtechs.net/photography/the-rise-of-profile-photos
+http://jrtechs.net/photography/photography-in-the-age-of-social-media
+http://jrtechs.net/programming/csci-344-final-review
+http://jrtechs.net/programming/multi-threaded-file-io
 http://jrtechs.net/programming/sorting-algorithms
 http://jrtechs.net/programming/knapsack-problem
 http://jrtechs.net/programming/cs-theory-exam-2-review
@@ -26,6 +48,8 @@ http://jrtechs.net/programming/everything-fibonacci
 http://jrtechs.net/programming/c-to-c++-tutorial
 http://jrtechs.net/programming/university-vs-teaching-yourself-programming
 http://jrtechs.net/programming/using-english-conventions-to-write-clean-code
+http://jrtechs.net/projects/diy-video-hosting-server
+http://jrtechs.net/projects/github-graphs-project
 http://jrtechs.net/projects/steam-friends-graph
 http://jrtechs.net/projects/musical-floppy-drive-build-log
 http://jrtechs.net/projects/musical-floppy-drives
@@ -34,8 +58,43 @@ http://jrtechs.net/projects/ackermann-function-written-in-java
 http://jrtechs.net/projects/batch-minecraft-launcher-with-auto-restart
 http://jrtechs.net/projects/time-lapse-programming-zombie-game
 http://jrtechs.net/projects/time-lapse-programming-pong
+http://jrtechs.net/web-development/pandoc-syntax-highlighting-with-prism
+http://jrtechs.net/web-development/node-rss-feed
 http://jrtechs.net/web-development/lazy-loading-youtube-videos
 http://jrtechs.net/web-development/node-website-optimization
 http://jrtechs.net/web-development/node-vs-php
 http://jrtechs.net/web-development/why-i-stopped-using-wordpress
 http://jrtechs.net/web-development/history-of-jrtechs
+http://jrtechs.net/hardware/creating-an-audio-switch
+http://jrtechs.net/hardware/2018-rochester-maker-faire
+http://jrtechs.net/hardware/ddr4-ram-introduction
+http://jrtechs.net/hardware/hard-drive-speeds
+http://jrtechs.net/other/2020-in-review
+http://jrtechs.net/other/coming-out-as-bisexual
+http://jrtechs.net/other/college-cookbook-part-3
+http://jrtechs.net/other/flirting-with-burnout-at-rit
+http://jrtechs.net/other/how-films-portray-control
+http://jrtechs.net/other/working-remote
+http://jrtechs.net/other/college-cookbook-part-2
+http://jrtechs.net/other/2019-in-review
+http://jrtechs.net/other/responsible-optimization
+http://jrtechs.net/other/college-cookbook
+http://jrtechs.net/other/2018-in-review
+http://jrtechs.net/other/morality-of-self-driving-cars
+http://jrtechs.net/other/my-college-essay
+http://jrtechs.net/other/deepfakes
+http://jrtechs.net/other/why-do-i-blog
+http://jrtechs.net/other/ways-to-avoid-getting-hacked
+http://jrtechs.net/other/should-you-run-a-server-on-desktop-hardware
+http://jrtechs.net/open-source/creating-a-dynamic-github-profile
+http://jrtechs.net/open-source/hosting-your-own-gitea-server
+http://jrtechs.net/open-source/community-architecture
+http://jrtechs.net/open-source/community-architecture-proposal
+http://jrtechs.net/open-source/hfoss-quiz-1
+http://jrtechs.net/open-source/ritlug-bugfix
+http://jrtechs.net/open-source/shallow-dive-into-open-cv
+http://jrtechs.net/open-source/foss-vs-floss
+http://jrtechs.net/open-source/jupyter-will-change-your-life
+http://jrtechs.net/open-source/towards-a-new-hacker-ethic
+http://jrtechs.net/open-source/teaching-ssh-through-a-ctf
+http://jrtechs.net/open-source/the-essential-vim-configuration
