+++ templates/admin/adminDownloads.html
@@ -1,7 +1,7 @@
 <div class="row">
 
     <!-- Add Download -->
-    <div class="col-md-6">
+    <div class="col-md-4">
         <div class="blogPost">
             <h1 class="text-center">Add Download</h1>
 
@@ -24,15 +24,16 @@
     </div>
 
     <!-- Downloads -->
-    <div class="col-md-6">
+    <div class="col-md-8">
         <div class='blogPost'>
             <h1 class="text-center">Downloads</h1>
             <div class="">
-                <table class="table table-striped">
+                <table class="table table-responsive table-striped">
                     <thead class="thead-dark">
                         <tr>
                             <td>Download Name</td>
                             <td>File</td>
+                            <td>URL</td>
                             <td>Download Count</td>
                             <td>Delete</td>
                         </tr>
@@ -43,6 +44,9 @@
                             <tr>
                                 <td>
                                     {download.name}
+                                </td>
+                                <td class="w-25">
+                                    https://jrtechs.net/downloads/{download.name}
                                 </td>
                                 <td>
                                     {download.file}
